Fixed bugs in textService
- using getSplit did not notify the preloaders that the text was loaded
- if an observable is used multiple times (for instance if it is
binded multiple times in the html), the preloaders would be notified
that the image is loaded for every usage.

Unluckily (or luckily depending on the pov), the app had 1 text
binded twice in the header
=> the preloader was notified twice for the same loaded text
and the app only had 1 call of getSplit
=> the preloader was not notified for precisely 1 text loaded
Hence, those two mistakes cancelled out. If the balance was incorrect,
the preloaders would just have kept loading and the page would have
been inaccessible, or a warning message would have occured since the
preloader service would have received to many loaded notifications. The
former happened when I tried to add a new API call for the alt text.

And that wasn't even the issue with the alt text, the real issue was
that the async pipe wasn't present, but it was what let me to do an
investigation and to find this bug.